Refactor connectivity to Display-mediated platform classes; isolate API 21+ on Android
Addresses review feedback on PR #5021:

1) Encapsulation: drop the IOImpl backdoor that exposed
   CodenameOneImplementation publicly. Instead each new API group has a
   narrow platform interface (WifiPlatform, WifiDirectPlatform,
   BonjourPlatform, UsbPlatform, NetworkTypePlatform) obtained from
   Display.getInstance().getXxxPlatform(). CodenameOneImplementation
   only carries small createXxxPlatform() factories; each port returns
   its concrete subclass, so the base implementation stays modular.

2) Old-device safety: AndroidConnectivity held static fields of type
   ConnectivityManager.NetworkCallback (API 21) which would fail
   class verification on KitKat (minSdkVersion 19). The Lollipop-only
   bits now live in *Lollipop helper classes that are only referenced
   from inside Build.VERSION.SDK_INT guards, so the KitKat verifier
   never loads them.

3) NetworkManager.isConnected(): cheap EDT-safe reachability check
   that returns false only when NETWORK_TYPE_NONE, avoiding the HTTP
   probe against autoDetectURL.

4) Documentation:
   - Drop the version "Codename One 8" mention.
   - Clarify JmDNS belongs in the application's executable-jar /
     simulator profile, not in the cn1app common pom (which would
     ship JmDNS to devices that already have native mDNS).
   - Replace the deinitializeImpl() lifecycle note with a battery-
     drain warning -- deinitializeImpl is internal API and shouldn't
     show up in developer docs.
   - Add a quick-check example for NetworkManager.isConnected().

5) SpotBugs exclusion comment: rewrite the rationale to explain why
   LI_LAZY_INIT_* doesn't apply (EDT-bound install pattern, no
   actual race) instead of just "matches existing convention".
