+++ Ports/JavaSE/src/com/codename1/impl/javase/JavaSEPort.java
@@ -12463,100 +12463,23 @@ public class JavaSEPort extends CodenameOneImplementation {
     }
 
     @Override
-    public int getCurrentNetworkType() {
-        return JavaSEConnectivity.getCurrentNetworkType();
+    protected com.codename1.io.wifi.WifiPlatform createWifiPlatform() {
+        return new com.codename1.impl.javase.connectivity.JavaSEWifiPlatform();
     }
 
     @Override
-    public boolean isWiFiInfoSupported() {
-        return JavaSEConnectivity.isWiFiInfoSupported();
+    protected com.codename1.io.wifi.WifiDirectPlatform createWifiDirectPlatform() {
+        return new com.codename1.impl.javase.connectivity.JavaSEWifiDirectPlatform();
     }
 
     @Override
-    public boolean isWiFiManagementSupported() {
-        return JavaSEConnectivity.isWiFiManagementSupported();
+    protected com.codename1.io.bonjour.BonjourPlatform createBonjourPlatform() {
+        return new com.codename1.impl.javase.connectivity.JavaSEBonjourPlatform();
     }
 
     @Override
-    public String getWiFiSSID() { return JavaSEConnectivity.getWiFiSSID(); }
-
-    @Override
-    public String getWiFiBSSID() { return JavaSEConnectivity.getWiFiBSSID(); }
-
-    @Override
-    public String getWiFiGateway() { return JavaSEConnectivity.getWiFiGateway(); }
-
-    @Override
-    public String getWiFiIp() { return JavaSEConnectivity.getWiFiIp(); }
-
-    @Override
-    public void scanWiFi(com.codename1.io.wifi.WiFiScanCallback cb) {
-        JavaSEConnectivity.scanWiFi(cb);
-    }
-
-    @Override
-    public void connectWiFi(String ssid, String password,
-                            com.codename1.io.wifi.WiFiSecurity security,
-                            com.codename1.io.wifi.WiFiConnectCallback cb) {
-        JavaSEConnectivity.connectWiFi(ssid, password, security, cb);
-    }
-
-    @Override
-    public void disconnectWiFi(String ssid) {
-        JavaSEConnectivity.disconnectWiFi(ssid);
-    }
-
-    @Override
-    public boolean isBonjourSupported() {
-        return JavaSEConnectivity.isBonjourSupported();
-    }
-
-    @Override
-    public Object startBonjourBrowse(String type,
-                                     com.codename1.io.bonjour.BonjourServiceListener l) {
-        return JavaSEConnectivity.startBonjourBrowse(type, l);
-    }
-
-    @Override
-    public void stopBonjourBrowse(Object handle) {
-        JavaSEConnectivity.stopBonjourBrowse(handle);
-    }
-
-    @Override
-    public Object startBonjourPublish(String name, String type, int port,
-                                      java.util.Map<String, String> txt) {
-        return JavaSEConnectivity.startBonjourPublish(name, type, port, txt);
-    }
-
-    @Override
-    public void stopBonjourPublish(Object handle) {
-        JavaSEConnectivity.stopBonjourPublish(handle);
-    }
-
-    @Override
-    public boolean isWiFiDirectSupported() {
-        return JavaSEConnectivity.isWiFiDirectSupported();
-    }
-
-    @Override
-    public void startWiFiDirectDiscovery(com.codename1.io.wifi.WiFiDirectListener l) {
-        JavaSEConnectivity.startWiFiDirectDiscovery(l);
-    }
-
-    @Override
-    public void stopWiFiDirectDiscovery() {
-        JavaSEConnectivity.stopWiFiDirectDiscovery();
-    }
-
-    @Override
-    public void connectWiFiDirect(com.codename1.io.wifi.WiFiDirectPeer peer,
-                                  com.codename1.io.wifi.WiFiConnectCallback cb) {
-        JavaSEConnectivity.connectWiFiDirect(peer, cb);
-    }
-
-    @Override
-    public void disconnectWiFiDirect() {
-        JavaSEConnectivity.disconnectWiFiDirect();
+    protected com.codename1.io.NetworkTypePlatform createNetworkTypePlatform() {
+        return new com.codename1.impl.javase.connectivity.JavaSENetworkTypePlatform();
     }
 
     @Override
